requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
aiohttp
anthropic
openai
python-dotenv
jinja2
requests